worker/requirements.txt
asyncpg
azure-storage-blob
azure-storage-queue
aiohttp